[feat] Parse leading N+M and N/M unit counts (e.g. 3+1RACL, 1/2 REBLOCH)
Make units a float to support fractional counts like 1/2. Parse leading
arithmetic prefixes: N+M sums (3+1 → 4) and N/M fractions (1/2 → 0.5).

+++ superslurp/compare/aggregate.py
@@ -77,7 +77,7 @@ def _build_observation(
     price_per_liter: float | None = None
     if volume_ml is not None and volume_ml > 0:
         price_per_liter = round((price / volume_ml) * 1000, 2)
-    unit_count: int = item.get("units") or 1
+    unit_count: float = item.get("units") or 1
     obs: dict[str, Any] = {
         "original_name": item["name"],
         "session_id": session_id,
